Add config override for profile page in navbar
This adds the config option of `show_profile_page_in_navbar` to allow the ability to disable the navbar entry. The main use case for this is if you want to override the filament views and provide a custom link to the profile page, without relying on the sidebar. For example, using the user dropdown.